Refactoring de ordered_map

+++ trunk/ordered_map/src/main.cc
@@ -1,4 +1,4 @@
-// Copyright 2011 Universidade Federal de Minas Gerais (UFMG)
+// Copyright 2014 Universidade Federal de Minas Gerais (UFMG)
 //
 // Lista sobre conjuntos associativos
 //
@@ -13,12 +13,13 @@
 // Você pode assumir que os arquivos contém apenas caracteres que são letras
 // (a-z e A-Z), números (0-9), e caracteres de pontuação (",", ".", "?", etc.).
 // Você pode assumir que o texto não tem acentos nem "ç".
-// Após ler cada palavra, você deve (i) transformar todas as letras maiúsculas em
-// minúsculas e (ii) apagar todos os caracteres que não são letras ou números. Por
-// exemplo, depois de ler "Guarda-Chuva?", você deve transformá-la em "guardachuva",s
-// antes de inseri-la no conjunto. Desta forma, a mesma palavra apresentada com
-// letras minúsculas ou maiúsculas, ou que estão adjacentes a pontuação, não serão
-// diferenciadas. Por exemplo, para o texto:
+// Após ler cada palavra, você deve (i) transformar todas as letras maiúsculas
+// em minúsculas e (ii) apagar todos os caracteres que não são letras ou
+// números. Por exemplo, depois de ler "Guarda-Chuva?", você deve
+// transformá-la em "guardachuva", antes de inseri-la no conjunto. Desta forma,
+// a mesma palavra apresentada com letras minúsculas ou maiúsculas,
+// ou que estão adjacentes a pontuação, não serão diferenciadas.
+// Por exemplo, para o texto:
 //
 // -------------------------
 // La vem o pato.
@@ -44,14 +45,11 @@
 // ver 1
 // -------------------------
 
-
-#include <cstdlib>
 #include <fstream>
 #include <iostream>
 #include <string>
-#include <ctype.h>
 
-#include "ordered_map/src/map.h"
+#include "src/map.h"
 
 using std::cin;
 using std::cout;
@@ -61,7 +59,7 @@ using std::string;
 
 string CleanString(string s) {
   string result;
-  for (int i = 0; i < (int) s.length(); i++) {
+  for (int i = 0; i < s.length(); i++) {
     if (isalnum(s[i])) {
       result.push_back(tolower(s[i]));
     }
@@ -77,7 +75,7 @@ int main() {
   while (fin >> str) {
     str = CleanString(str);
     word = words.find(str);
-    if (word == words.end() ) {
+    if (word == words.end()) {
       words.insert(str, 1);
     } else {
       words[str] = words[str] + 1;
